Move bundle price selector as a standalone component.
