+++ client/a8c-for-agencies/sections/marketplace/products-overview-v2/product-action-panel/style.scss
@@ -33,37 +33,3 @@
 		flex-grow: 1;
 	}
 }
-
-.bundle-price-selector {
-	display: flex;
-	flex-direction: row;
-	gap: 8px;
-	align-items: center;
-}
-
-.bundle-price-selector__label {
-	@include a4a-font-heading-sm;
-	text-transform: uppercase;
-	display: none;
-
-	@include break-medium {
-		display: block;
-	}
-}
-
-.bundle-price-selector__menu-button {
-	flex-grow: 1;
-}
-
-.bundle-price-selector__popover {
-	margin-inline-start: -135px;
-
-	.popover__menu .popover__menu-item {
-		display: block;
-
-		&.is-selected {
-			background: var( --color-neutral-5 );
-			color: var( --color-text );
-		}
-	}
-}
